fix: moves cleanup utility
