+++ samples/test/clean.js
