+++ samples/test/get_realm.test.js
@@ -15,7 +15,7 @@
 'use strict';
 
 const {assert} = require('chai');
-const cleanup = require('./utils/clean.js');
+const cleanup = require('./clean.js');
 const {describe, it, before, after} = require('mocha');
 const {RealmsServiceClient} = require('@google-cloud/game-servers');
